flask-app/requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
Flask==3.1.3
flask-cors==6.0.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyYAML==6.0.3
requests==2.32.5
urllib3==2.6.3
Werkzeug==3.1.6
